* master: (41 commits)
  Added Response helper for 'attachments' [#164 state:resolved]
  Removed deprecated support for delivered_emails [#152 state:resolved]
  Removed deprecated support for droute_url [#106 state:resolved]
  Removed deprecated support for 'top' level ApplicationHelper [#93 state:resolved]
  updated changelog
  Rake task to convert app config yamls to configatron [#167 state:resolved]
  Fixed ERB errors not being reported correctly. [#168 state:resolved]
  Do not cache classes in development
  Raise errors on boot loading.
  Updated to configatron 2.1.1
  Updated CHANGELOG
  minor update to gem_freezer environment handler
  Finished conversion to configatron 2.0.0
  Working on converting to configatron 2.x
  Freeze Gems [#119 state:resolved]
  freezing gem: solved loading issues
  more gem_freezing updates
  Fixed configs getting loaded twice.
  Removed unnecessary cachetastic require.
  Set the correct version of cachetastic
  ...

+++ lib/mack/runner_helpers/session.rb
@@ -6,7 +6,7 @@ module Mack
       attr_accessor :sess_id
       
       def start(request, response, cookies)
-        if app_config.mack.use_sessions
+        if configatron.mack.use_sessions
           self.sess_id = retrieve_session_id(request, response, cookies)
           unless self.sess_id
             self.sess_id = create_new_session(request, response, cookies)
@@ -23,20 +23,22 @@ module Mack
       end
       
       def complete(request, response, cookies)
-        unless response.redirection?
-          request.session.delete(:tell)
+        if configatron.mack.use_sessions
+          unless response.redirection?
+            request.session.delete(:tell)
+          end
+          Mack::SessionStore.set(request.session.id, request, response, cookies)
         end
-        Mack::SessionStore.set(request.session.id, request, response, cookies) if app_config.mack.use_sessions
       end
       
       private
       def retrieve_session_id(request, response, cookies)
-        cookies[app_config.mack.session_id]
+        cookies[configatron.mack.session_id]
       end
       
       def create_new_session(request, response, cookies)
         id = String.randomize(40).downcase
-        cookies[app_config.mack.session_id] = {:value => id, :expires => nil}
+        cookies[configatron.mack.session_id] = {:value => id, :expires => nil}
         sess = Mack::Session.new(id)
         request.session = sess
         Mack::SessionStore.set(request.session.id, request, response, cookies)
